+++ scripts/constants.ts
@@ -1,4 +1,0 @@
-const path = require("path");
-
-export const rebuildPath = path.resolve(__dirname, "../public/rebuild");
-export const distPath = path.resolve(__dirname, "../.output/public");