+++ scripts/generate.ts
@@ -1,9 +1,6 @@
 import fs from "fs";
-import { resolve } from "path";
 import { getNowDayjsString } from "../utils/common";
-import genRss from "./rss";
-import { distPath } from "./constants";
-import { nbLog, runCmd } from ".";
+import { genRss, getDistPath, getRebuildPath, nbLog, runCmd } from "./utils";
 
 export default async function () {
   await runCmd("nuxt generate");
@@ -13,15 +10,15 @@ export default async function () {
 
 function generateSiteMap () {
   nbLog("sitemap");
-  fs.writeFileSync(resolve(distPath, "sitemap.xml"),
-    genRss(JSON.parse(fs.readFileSync(resolve(__dirname, "../public/rebuild/json/articles.json")).toString())));
+  fs.writeFileSync(getDistPath("sitemap.xml"),
+    genRss(JSON.parse(fs.readFileSync(getRebuildPath("json", "articles.json")).toString())));
 }
 
 function generateTimestamp () {
   nbLog("inject timestamp");
   const timestamp = getNowDayjsString();
-  fs.writeFileSync(resolve(distPath, "timestamp.txt"), timestamp);
-  const aboutHtmlPath = resolve(distPath, "about/index.html");
+  fs.writeFileSync(getDistPath("timestamp.txt"), timestamp);
+  const aboutHtmlPath = getDistPath("about", "index.html");
   if (fs.existsSync(aboutHtmlPath)) {
     fs.writeFileSync(
       aboutHtmlPath,